+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added support to manage the tags in the create or edit activity dialog
 - Added the tags to the admin control panel
 
 ### Changed